--- src/components/map-2/style.module.scss
@@ -1,6 +1,0 @@
-.map {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  z-index: 1;
-}
