+++ src/components/map-container/index.js
@@ -1,0 +1,18 @@
+import { connect } from 'react-redux';
+import { setViewport } from 'modules/map/actions';
+import { mapStyle } from 'modules/map-styles/selectors';
+
+import Component from './component';
+
+const mapStateToProps = state => ({
+  ...state.map,
+  mapStyle: mapStyle(state),
+  mapboxApiAccessToken: process.env.REACT_APP_MAPBOX_ACCESS_TOKEN,
+  isCollapse: state.layers.isCollapsed
+});
+
+const mapDispatchToProps = {
+  setViewport
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(Component);
