+++ src/components/map-container/style.module.scss
@@ -1,0 +1,49 @@
+@import 'styles/vars';
+
+$space-around-map: 30px;
+$space-top-map-mobile: 5px;
+$space-left-map-mobile: 20px;
+
+.map {
+  width: 100%;
+  height: 100%;
+  position: relative;
+}
+
+  .navigation {
+    position: absolute;
+    right: $space-around-map;
+    top: $space-around-map;
+  }
+
+  .legend {
+    position: absolute;
+    right: $space-around-map;
+    bottom: $space-around-map;
+    z-index: 1;
+
+    @media screen and (max-width: map-get($breakpoints, md)) {
+      top: $space-top-map-mobile;
+      left: $space-left-map-mobile;
+      &.expanded {
+        display: flex;
+        flex-direction: row-reverse;
+        bottom: inherit;
+        right: 25px;
+      }
+    }
+
+    .tooltip {
+      display: flex;
+      flex-direction: column;
+    }
+  }
+
+  // Mapbox override
+  :global(.mapboxgl-ctrl-group:not(:empty)) {
+    box-shadow: 0 4px 12px 0 rgba(168, 168, 168, 0.25);
+  }
+
+  :global(.mapboxgl-ctrl-group > button) {
+    background: white;
+  }
